+++ src/main/java/com/hoonterpark/concertmanager/infrastructure/UserRepositoryImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 
+
 @Repository
 @RequiredArgsConstructor
 public class UserRepositoryImpl implements UserRepository {
