+++ src/test/java/com/hoonterpark/concertmanager/domain/ConcertServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.hoonterpark.concertmanager;
+package com.hoonterpark.concertmanager.domain;
 
 import com.hoonterpark.concertmanager.domain.entity.ConcertEntity;
 import com.hoonterpark.concertmanager.domain.entity.ConcertScheduleEntity;
