+++ src/test/java/com/hoonterpark/concertmanager/domain/SeatServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.hoonterpark.concertmanager;
+package com.hoonterpark.concertmanager.domain;
 
 import com.hoonterpark.concertmanager.domain.entity.SeatEntity;
 import com.hoonterpark.concertmanager.domain.enums.SeatStatus;
