+++ src/test/java/com/hoonterpark/concertmanager/infrastructure/TokenRepositoryIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.hoonterpark.concertmanager;
+package com.hoonterpark.concertmanager.infrastructure;
 
 
 import com.hoonterpark.concertmanager.domain.entity.TokenEntity;
